refactor(team): extract options + preflight + escalation (C17)
Three more semantic divisions extracted from launcher.go.

New files:

* launcher_options.go (~100 lines, new) — the user-facing config
  knob surface: SetUnsafe, SetOpusCEO, SetFocusMode, SetNoOpen,
  SetOneOnOne, plus the small derived predicates that read those
  knobs back (isOneOnOne, oneOnOneAgent, usesCodexRuntime,
  usesOpencodeRuntime, UsesTmuxRuntime), the exported accessors the
  channel TUI / cmd/wuphf use (BrokerToken, OneOnOneAgent), and
  isBlankSlateLaunchSlug. The Launcher's "what mode am I in?"
  surface — separate from construction (NewLauncher), orchestration
  (Launch), and sub-type wiring.

* launcher_preflight.go (~70 lines, new) — Preflight (claude /
  codex / opencode binary check) + checkGHCapability (the gh CLI
  status probe used to surface a setup-incomplete advisory). Pairs
  with PreflightWeb in launcher_web.go which is the browser-mode
  equivalent.

* escalation.go (~40 lines, new) — postEscalation, the broker-write
  helper that surfaces agent-stuck / max-retries / generic
  escalations into #general as Slack-style heads-ups. Pure broker
  passthrough plus the self-healing kick.

launcher.go: 728 -> 579 lines (-149).

Cumulative since main: 4998 -> 579 = -4419 lines (-88.4%).

Stacked on refactor/launcher-c16-launcher-types (C16).
